move parse_yaml_msg_id_file() to uorb_rtps_classifier.py
